// src/features/database/types.ts
export type CarData = {
  car: string;
  height: number | null;
  fAxleOffset: number | null;
  rAxleOffset: number | null;
  wheelbase: number | null;
  fTrackWidth: number | null;
  rTrackWidth: number | null;
  avTrackWidth: number | null;
  gears: number | null;
  shiftTime: number | null;
  weight: number | null;
  stockCx: number | null;
  stockSx: number | null;
  stockDrag: number | null;
  stage12Cx: number | null;
  stage12Sx: number | null;
  stage12Drag: number | null;
  stage34Cx: number | null;
  stage34Sx: number | null;
  stage34Drag: number | null;
  bodyPosX: number | null;
  bodyPosY: number | null;
  bodyPosZ: number | null;
  powerHp: number | null;
  massKg: number | null;
  turboPress: number | null;
  curveFallRpm: number | null;
  revLimiter: number | null;
  inertiaRatio: number | null;
  enginePosX: number | null;
  enginePosY: number | null;
  enginePosZ: number | null;
};